backend/requirements.txt
# ── Web framework ─────────────────────────────────────────────────────────────
fastapi==0.115.5
uvicorn[standard]==0.32.1
python-multipart==0.0.12

# ── Database ───────────────────────────────────────────────────────────────────
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
psycopg2-binary==2.9.10
alembic==1.14.0
geoalchemy2==0.16.0

# ── Cache / Redis ──────────────────────────────────────────────────────────────
redis[hiredis]==5.2.0

# ── Temporal ──────────────────────────────────────────────────────────────────
temporalio==1.7.1

# ── Auth / crypto ─────────────────────────────────────────────────────────────
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
cryptography==43.0.3

# ── Validation / settings ─────────────────────────────────────────────────────
pydantic==2.9.2
pydantic-settings==2.6.1
email-validator==2.2.0

# ── HTTP client ───────────────────────────────────────────────────────────────
httpx==0.28.0

# ── Cloud storage (R2 / S3-compatible) ────────────────────────────────────────
boto3==1.35.80

# ── Image processing ──────────────────────────────────────────────────────────
Pillow==11.0.0
piexif==1.1.3

# ── Fuzzy name matching ───────────────────────────────────────────────────────
rapidfuzz==3.10.1

# ── Utilities ─────────────────────────────────────────────────────────────────
python-dotenv==1.0.1
structlog==24.4.0
orjson==3.10.12
phonenumbers==8.13.49
pyotp==2.9.0

# ── Testing ───────────────────────────────────────────────────────────────────
pytest==8.3.4
pytest-asyncio==0.24.0
pytest-mock==3.14.0
httpx==0.28.0
factory-boy==3.3.1
